Adjust stan::length(x) to handle Eigen::MatrixBase

+++ src/stan/math/prim/mat/fun/qr_R.hpp
@@ -14,10 +14,7 @@ namespace stan {
     qr_R(const Eigen::Matrix<T,Eigen::Dynamic,Eigen::Dynamic>& m) {
       typedef Eigen::Matrix<T,Eigen::Dynamic,Eigen::Dynamic> matrix_t;
       stan::math::check_nonzero_size("qr_R", "m", m);
-      stan::math::check_greater_or_equal("qr_R",
-                                                   "m.rows()",
-                                                   static_cast<size_t>(m.rows()),
-                                                   static_cast<size_t>(m.cols()));
+      stan::math::check_greater_or_equal("qr_R", "m.rows()", m.rows(), m.cols());
       Eigen::HouseholderQR<matrix_t> qr(m.rows(), m.cols());
       qr.compute(m);
       matrix_t R = qr.matrixQR().topLeftCorner(m.rows(),m.cols());
